requirements.txt
timm==0.4.12
transformers==4.17.0
fairscale==0.4.4
requests
pycocoevalcap
torch
torchvision
Pillow
fastapi[standard]